requirements.txt
gradio
pandas
plotly
tabulate
langchain
langchain-community
google-generativeai
